BackEnd/requirements.txt
django
djangorestframework

djangorestframework-simplejwt
djoser
django-cors-headers
pytz
psycopg2-binary

social-auth-app-django
numpy